+++ Frontend/src/components/SearchForm.js
@@ -42,13 +42,43 @@ class SearchForm extends Component{
       handleDelete = (index) => {
         const newArr = [...this.state.tasks];
         // console.log(newArr[index],index);
-        let task = newArr[index];
+        let pos = -1;
+        for( let i =0 ; i< newArr.length;i++ ){
+          if( newArr[i].id === index){
+            pos = i;
+            break;
+          }
+        }
+        if( pos == -1) return;
+        let task = newArr[pos];
         console.log(task);
         axios.delete(`http://localhost:3001/todos`,  { data: task })
         .then(res => {
           console.log("In Frontend",res.data);
         });
-        newArr.splice(index, 1);
+        newArr.splice(pos, 1);
+        this.setState({tasks: newArr});
+      }
+      handleUpdate = (task) => {
+        const newArr = [...this.state.tasks];
+    
+        let pos = -1;
+        for( let i =0 ; i< newArr.length; i++ ){
+          if( newArr[i].id === task.id){
+            pos = i;
+            break;
+          }
+        }
+        if( pos == -1) return;
+        newArr[pos].name = task.name;
+        newArr[pos].startDate = task.startDate;
+        newArr[pos].endDate= task.endDate;
+        let updateTask = newArr[pos];
+        console.log(updateTask);
+        axios.put(`http://localhost:3001/todos`,  { data: updateTask })
+        .then(res => {
+          console.log("Updating In Frontend",res.data);
+        });
         this.setState({tasks: newArr});
       }
       handleInputChange = (event) => {
@@ -108,7 +138,8 @@ class SearchForm extends Component{
             />
             <button class='button1'>Submit</button>
           </form>
-          <TodoList tasks={this.state.tasks} onDelete={this.handleDelete}  />
+          <TodoList tasks={this.state.tasks} onDelete={this.handleDelete} 
+             onUpdate={this.handleUpdate} />
          
           </div> 
         );
